repaired uninitialized variable

+++ src/math_and_physics/vector_operations.h
@@ -23,7 +23,7 @@ void Vect_write_three_values(Vector_t *vector, double value_1, double value_2, d
 void Vect_times_const(Vector_t *vector, double constant, Vector_t *result_vector);
 void Vect_copy_from_to(Vector_t *vector_from, Vector_t *vector_to);
 void Vect_cross_multiply(Vector_t *vector_1, Vector_t *vector_2, Vector_t *resultvector);
-void Vect_sum_up_list_of_vectors(Vector_t *vectorlist, Vector_t *sum_vector, uint8_t listlength);
+void Vect_sum_up_list_of_vectors(Vector_t *vectorlist, Vector_t *sum_vector, uint32_t listlength);
 void Vect_set_vectorlist_to_value(Vector_t vectorlist[], uint32_t listlength, double value);
 double *Vect_pointer_to_index(Vector_t *vect, uint8_t index);
